Update 2.6.x to r64522 svn merge -r64473:64522 https://source.sakaiproject.org/contrib/assignment2/trunk --- Merging r64474 through r64522 into '.': U    impl/src/java/org/sakaiproject/assignment2/logic/impl/ExternalContentReviewLogicImpl.java U    api/src/java/org/sakaiproject/assignment2/model/constants/AssignmentConstants.java U    tool/src/java/org/sakaiproject/assignment2/tool/LocalTurnitinLogic.java U    tool/src/java/org/sakaiproject/assignment2/tool/producers/AssignmentProducer.java U    tool/src/java/org/sakaiproject/assignment2/tool/producers/ViewSubmissionsProducer.java U    tool/src/java/org/sakaiproject/assignment2/tool/entity/Assignment2SubmissionEntityProvider.java U    tool/src/webapp/WEB-INF/applicationContext.xml U    tool/src/webapp/WEB-INF/requestContext.xml U    tool/src/webapp/content/js/submissionview.js U    tool/src/webapp/component-templates/AttachmentEvolver.html ------------------------------------------------------------------------ r64474 | swgithen@mtu.edu | 2009-11-05 08:54:29 -0500 (Thu, 05 Nov 2009) | 1 line
ASNN-516 ONC-2097 Fixing message key code for SUBMITTION_ERROR_RETRY_EXCEEDED_CODE
------------------------------------------------------------------------
r64504 | wagnermr@iupui.edu | 2009-11-06 12:00:04 -0500 (Fri, 06 Nov 2009) | 3 lines

ONC-2092
https://uisapp2.iu.edu/jira-prd/browse/ONC-2092
Submit paper to IU repository should be default setting
------------------------------------------------------------------------
r64519 | swgithen@mtu.edu | 2009-11-07 21:52:14 -0500 (Sat, 07 Nov 2009) | 1 line

ASNN-539 ONC-1846 Added ability to save the sorting and pagesize for the Submissions View on a given site. This is going in session state, which is not awesome, but it's not much data. Just a couple strings. All the data for the table is fresh from the DB for every table page/resize. When the user changes a table by sorting, or resizing, the EntityBroker feed looks to see if a tool placementId was passed in on the call. If it was, it saves the new sort/size information in the tool session. If there is no placement ID it doesn't, which allows the JSON feed to still be useful for new applications that are not bound by tool placements. Each the View Submission Page is reloaded, the producer checks to see if this information has been saved in the tool session, and if it has uses it for the initial table load. 
------------------------------------------------------------------------
r64522 | swgithen@mtu.edu | 2009-11-08 19:23:05 -0500 (Sun, 08 Nov 2009) | 1 line

ASNN-571 ONC-2082 Removing 2 hardcoded URLs that was in a hidden div pointing at localhost:8080
------------------------------------------------------------------------


git-svn-id: https://source.sakaiproject.org/contrib/assignment2/branches/sakai_2-6-x@64545 fdecad78-55fc-0310-b1b2-d7d25cf747c9