Update 2.6.x to r67319 svn merge -r67308:67319 https://source.sakaiproject.org/contrib/assignment2/trunk --- Merging r67309 through r67319 into '.': U    impl/src/java/org/sakaiproject/assignment2/logic/impl/AssignmentLogicImpl.java U    tool/src/java/org/sakaiproject/assignment2/tool/beans/SessionCache.java U    tool/src/java/org/sakaiproject/assignment2/tool/beans/AssignmentAuthoringBean.java U    tool/src/java/org/sakaiproject/assignment2/tool/producers/GradeProducer.java U    tool/src/java/org/sakaiproject/assignment2/tool/entity/Assignment2SubmissionEntityProvider.java U    tool/src/webapp/WEB-INF/applicationContext.xml ------------------------------------------------------------------------ r67309 | swgithen@mtu.edu | 2010-04-23 16:01:21 -0400 (Fri, 23 Apr 2010) | 1 line
ASNN-529 ONC-2070 Refactoring things a bit so that the studentID list can be regenerated with just teh sort options stored in Session in case the cache expires while instructors are going through a list of students. Involved restructuring the entity provider to allow signatures that took explicit options rather than depending on request parameters. As well as adding filter group tothe table sort options that are stored so it can be used when regenerting the list.
------------------------------------------------------------------------
r67319 | wagnermr@iupui.edu | 2010-04-26 09:09:18 -0400 (Mon, 26 Apr 2010) | 2 lines

ASNN-652
If initial link to TII fails when you are creating an assignment, do not mark as TII Enabled
------------------------------------------------------------------------


git-svn-id: https://source.sakaiproject.org/contrib/assignment2/branches/sakai_2-6-x@67320 fdecad78-55fc-0310-b1b2-d7d25cf747c9