ASNN-722 Most of the work for more informative error messages when turnitin submissions to not get submitted correctly. This goes hand in hand with some recent API additions in ContentReview for fetching translated error messages.
Because the number of properties being attached to assignments and
submissions was getting confusing with so many constants, we've
created a model object to attach as a single property. This 
contains a ContentReviewItem and adds a couple calculated properties.

This then trickles through all the UI rendering pieces and finally
in ExternalContentReviewLogicImpl we check the status of the review
item, and if it's in error we go to the ContentReviewService to 
get the detailed error message based on the Error Code from Turnitin.


git-svn-id: https://source.sakaiproject.org/contrib/assignment2/trunk@71207 fdecad78-55fc-0310-b1b2-d7d25cf747c9